chore: separate cell operation from row mod
