+++ frontend/rust-lib/flowy-grid/src/services/field/type_options/util/cell_data_util.rs
@@ -1,4 +1,4 @@
-use crate::services::row::AnyCellData;
+use crate::services::cell::AnyCellData;
 use flowy_grid_data_model::revision::CellRevision;
 use std::str::FromStr;
