+++ frontend/rust-lib/flowy-grid/src/services/mod.rs
@@ -2,6 +2,7 @@ mod util;
 
 mod block_manager;
 pub mod block_revision_editor;
+pub mod cell;
 pub mod field;
 mod filter;
 pub mod grid_editor;
