+++ frontend/rust-lib/flowy-grid/src/services/row/mod.rs
@@ -1,7 +1,5 @@
-mod cell_data_operation;
 mod row_builder;
 mod row_loader;
 
-pub use cell_data_operation::*;
 pub use row_builder::*;
 pub(crate) use row_loader::*;
